# airbyte-api/src/main/openapi/api.yaml
openapi: 3.0.0
info:
  title: Airbyte Public API
  version: 1.0.0
servers:
  - url: "http://localhost:8006/v1"
    description: Airbyte API v1
paths:
...
  /jobs:
    get:
      tags:
        - Jobs
      parameters:
        - name: connectionId
          description: Filter the Jobs by connectionId.
          schema:
            format: UUID
            type: string
          in: query
          required: false
        - name: limit
          description: Set the limit on the number of Jobs returned. The default is 20 Jobs.
          schema:
            format: int32
            default: 20
            maximum: 100
            minimum: 1
            type: integer
          in: query
        - name: offset
          description: Set the offset to start at when returning Jobs. The default is 0.
          schema:
            format: int32
            default: 0
            minimum: 0
            type: integer
          in: query
        - name: jobType
          description: Filter the Jobs by jobType.
          schema:
            $ref: "#/components/schemas/JobTypeEnum"
          in: query
        - name: workspaceIds
          description: The UUIDs of the workspaces you wish to list jobs for. Empty list will retrieve all allowed workspaces.
          schema:
            type: array
            items:
              format: uuid
              type: string
          in: query
          required: false
        - name: status
          description: The Job status you want to filter by
          schema:
            $ref: "#/components/schemas/JobStatusEnum"
          in: query
          required: false
        - name: createdAtStart
          description: The start date to filter by
          schema:
            type: string
            format: date-time
          in: query
          required: false
          example: 2023-06-22T16:15:00Z
        - name: createdAtEnd
          description: The end date to filter by
          schema:
            type: string
            format: date-time
          in: query
          required: false
          example: 2023-06-22T16:15:00Z
        - name: updatedAtStart
          description: The start date to filter by
          schema:
            type: string
            format: date-time
          example: 2023-06-22T16:15:00Z
          in: query
          required: false
        - name: updatedAtEnd
          description: The end date to filter by
          schema:
            type: string
            format: date-time
          in: query
          required: false
          example: 2023-06-22T16:15:00Z
        - name: orderBy
          description: The field and method to use for ordering
          schema:
            type: string
            pattern: \w+|(ASC|DESC)
          in: query
          required: false
          example: updatedAt|DESC
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsResponse"
              examples:
                Job List Response Example:
                  value:
                    next: https://api.airbyte.com/v1/jobs?limit=5&offset=10
                    previous: https://api.airbyte.com/v1/jobs?limit=5&offset=0
                    data:
                      - id: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                        status: running
                        jobType: sync
          description: List all the Jobs by connectionId.
        "403":
          description: Not allowed
      operationId: listJobs
      summary: List Jobs by sync type
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreateRequest"
            examples:
              Job Creation Request Example:
                value:
                  connectionId: e735894a-e773-4938-969f-45f53957b75b
                  jobType: sync
        required: true
      tags:
        - Jobs
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
              examples:
                Job Creation Response Example:
                  value:
                    jobId: 1234
                    status: running
                    jobType: sync
          description: "Kicks off a new Job based on the JobType. The connectionId is the resource that Job will be run for."
        "400":
          description: Invalid data
        "403":
          description: Not allowed
      operationId: createJob
      summary: Trigger a sync or reset job of a connection
  /jobs/{jobId}:
    get:
      tags:
        - Jobs
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
              examples:
                Job Get Response Example:
                  value:
                    id: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    status: running
                    jobType: sync
          description: Get a Job by the id in the path.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: getJob
      summary: Get Job status and details
    delete:
      tags:
        - Jobs
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
          description: Cancel a Job.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: cancelJob
      summary: Cancel a running Job
    parameters:
      - name: jobId
        schema:
          format: int64
          type: integer
        in: path
        required: true
...
  /connections/{connectionId}:
    get:
      tags:
        - Connections
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionResponse"
              examples:
                Connection Get Response Example:
                  value:
                    workspaceId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: Postgres To Snowflake
                    sourceId: 9924bcd0-99be-453d-ba47-c2c9766f7da5
                    destinationId: 744cc0ed-7f05-4949-9e60-2a814f90c035
          description: Get a Connection by the id in the path.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: getConnection
      summary: Get Connection details
    patch:
      tags:
        - Connections
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionPatchRequest"
            examples:
              Connection Update Request Example:
                value:
                  sourceId: 95e66a59-8045-4307-9678-63bc3c9b8c93
                  destinationId: e478de0d-a3a0-475c-b019-25f7dd29e281
                  name: Postgres-to-Bigquery
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionResponse"
              examples:
                Connection Get Response Example:
                  value:
                    workspaceId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: Postgres To Snowflake
                    sourceId: 9924bcd0-99be-453d-ba47-c2c9766f7da5
                    destinationId: 744cc0ed-7f05-4949-9e60-2a814f90c035
          description: Update a Connection by the id in the path.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: patchConnection
      summary: Update Connection details
    delete:
      tags:
        - Connections
      responses:
        "204":
          description: The resource was deleted successfully
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: deleteConnection
      summary: Delete a Connection
    parameters:
      - name: connectionId
        schema:
          format: UUID
          type: string
        in: path
        required: true
...
components:
...
  schemas:
...
    JobResponse:
      title: Root Type for JobResponse
      description: Provides details of a single job.
      required:
        - jobId
        - status
        - jobType
        - startTime
        - connectionId
      type: object
      properties:
        jobId:
          format: int64
          type: integer
        status:
          $ref: "#/components/schemas/JobStatusEnum"
        jobType:
          $ref: "#/components/schemas/JobTypeEnum"
        startTime:
          type: string
        connectionId:
          format: UUID
          type: string
        lastUpdatedAt:
          type: string
        duration:
          description: "Duration of a sync in ISO_8601 format"
          type: string
        bytesSynced:
          format: int64
          type: integer
        rowsSynced:
          format: int64
          type: integer
      example:
        id: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
        status: running
        jobType: sync
        startTime: 2023-03-25T01:30:50Z
        duration: PT8H6M12S
    JobsResponse:
      title: Root Type for JobsResponse
      description: ""
      required:
        - data
      type: object
      properties:
        previous:
          type: string
        next:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/JobResponse"
      example:
        next: https://api.airbyte.com/v1/jobs?limit=5&offset=10
        previous: https://api.airbyte.com/v1/jobs?limit=5&offset=0
        data:
          - id: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
            status: running
            jobType: sync
            startTime: 2023-03-25T01:30:50Z
...
    ConnectionPatchRequest:
      type: object
      properties:
        name:
          description: Optional name of the connection
          type: string
        configurations:
          $ref: "#/components/schemas/StreamConfigurations"
        schedule:
          $ref: "#/components/schemas/ConnectionSchedule"
        dataResidency:
          $ref: "#/components/schemas/GeographyEnumNoDefault"
        namespaceDefinition:
          $ref: "#/components/schemas/NamespaceDefinitionEnumNoDefault"
        namespaceFormat:
          type: string
          description: Used when namespaceDefinition is 'custom_format'. If blank then behaves like namespaceDefinition = 'destination'. If "${SOURCE_NAMESPACE}" then behaves like namespaceDefinition = 'source'.
          default: null
          example: "${SOURCE_NAMESPACE}"
        prefix:
          type: string
          description: Prefix that will be prepended to the name of each stream when it is written to the destination (ex. “airbyte_” causes “projects” => “airbyte_projects”).
        nonBreakingSchemaUpdatesBehavior:
          $ref: "#/components/schemas/NonBreakingSchemaUpdatesBehaviorEnumNoDefault"
        status:
          $ref: "#/components/schemas/ConnectionStatusEnum"
    JobCreateRequest:
      title: Root Type for JobCreate
      description: Creates a new Job from the configuration provided in the request body.
      required:
        - jobType
        - connectionId
      type: object
      properties:
        connectionId:
          format: UUID
          type: string
        jobType:
          $ref: "#/components/schemas/JobTypeEnum"
      example:
        connectionId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
        jobType: sync
    JobStatusEnum:
      enum:
        - pending
        - running
        - incomplete
        - failed
        - succeeded
        - cancelled
      type: string
    JobTypeEnum:
      description: Enum that describes the different types of jobs that the platform runs.
      enum:
        - sync
        - reset
      type: string
...
    ConnectionResponse:
      title: Root Type for ConnectionResponse
      description: Provides details of a single connection.
      type: object
      required:
        - connectionId
        - name
        - sourceId
        - destinationId
        - workspaceId
        - status
        - schedule
        - dataResidency
        - configurations
      properties:
        connectionId:
          format: UUID
          type: string
        name:
          type: string
        sourceId:
          format: UUID
          type: string
        destinationId:
          format: UUID
          type: string
        workspaceId:
          format: UUID
          type: string
        status:
          $ref: "#/components/schemas/ConnectionStatusEnum"
        schedule:
          $ref: "#/components/schemas/ConnectionScheduleResponse"
        dataResidency:
          $ref: "#/components/schemas/GeographyEnum"
        nonBreakingSchemaUpdatesBehavior:
          $ref: "#/components/schemas/NonBreakingSchemaUpdatesBehaviorEnum"
        namespaceDefinition:
          $ref: "#/components/schemas/NamespaceDefinitionEnum"
        namespaceFormat:
          type: string
        prefix:
          type: string
        configurations:
          $ref: "#/components/schemas/StreamConfigurations"
    ConnectionSchedule:
      description: schedule for when the the connection should run, per the schedule type
      type: object
      required:
        - scheduleType
      properties:
        scheduleType:
          $ref: "#/components/schemas/ScheduleTypeEnum"
        cronExpression:
          type: string
    ScheduleTypeEnum:
      type: string
      enum:
        - manual
        - cron
    ConnectionScheduleResponse:
      description: schedule for when the the connection should run, per the schedule type
      type: object
      required:
        - scheduleType
      properties:
        scheduleType:
          $ref: "#/components/schemas/ScheduleTypeWithBasicEnum"
        cronExpression:
          type: string
        basicTiming:
          type: string
    ScheduleTypeWithBasicEnum:
      type: string
      enum:
        - manual
        - cron
        - basic
    GeographyEnum:
      type: string
      enum:
        - auto
        - us
        - eu
      default: "auto"
    GeographyEnumNoDefault:
      type: string
      enum:
        - auto
        - us
        - eu
    ConnectionStatusEnum:
      type: string
      enum:
        - active
        - inactive
        - deprecated
    NamespaceDefinitionEnum:
      type: string
      description: Define the location where the data will be stored in the destination
      enum:
        - source
        - destination
        - custom_format
      default: "destination"
    NonBreakingSchemaUpdatesBehaviorEnum:
      type: string
      description: Set how Airbyte handles syncs when it detects a non-breaking schema change in the source
      enum:
        - ignore # do nothing if we detect a schema change
        - disable_connection # disable the connection, pausing the sync
        - propagate_columns # Auto propagate schema change for existing streams
        - propagate_fully # Auto propagate schema change for all streams including the new ones
      default: "ignore"
    NamespaceDefinitionEnumNoDefault:
      type: string
      description: Define the location where the data will be stored in the destination
      enum:
        - source
        - destination
        - custom_format
    NonBreakingSchemaUpdatesBehaviorEnumNoDefault:
      type: string
      description: Set how Airbyte handles syncs when it detects a non-breaking schema change in the source
      enum:
        - ignore # do nothing if we detect a schema change
        - disable_connection # disable the connection, pausing the sync
        - propagate_columns # Auto propagate schema change for existing streams
        - propagate_fully # Auto propagate schema change for all streams including the new ones
...
    StreamConfiguration:
      description: Configurations for a single stream.
      type: object
      required:
        - name
      properties:
        name:
          type: string
        syncMode:
          $ref: "#/components/schemas/ConnectionSyncModeEnum"
        cursorField:
          description: Path to the field that will be used to determine if a record is new or modified since the last sync. This field is REQUIRED if `sync_mode` is `incremental` unless there is a default.
          type: array
          items:
            type: string
        primaryKey:
          description: Paths to the fields that will be used as primary key. This field is REQUIRED if `destination_sync_mode` is `*_dedup` unless it is already supplied by the source schema.
          type: array
          items:
            type: array
            items:
              type: string
    StreamConfigurations:
      description: A list of configured stream options for a connection.
      type: object
      properties:
        streams:
          type: array
          items:
            $ref: "#/components/schemas/StreamConfiguration"
...
    ConnectionSyncModeEnum:
      enum:
        - full_refresh_overwrite
        - full_refresh_append
        - incremental_append
        - incremental_deduped_history
...